requirements.txt
# requirements.txt
pybit>=5.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
websocket-client>=1.6.0